refactor(element3): change import Tag component path
