+++ packages/element3/packages/cascader/Cascader.vue
@@ -139,7 +139,7 @@ import { useEmitter } from '../../src/use/emitter'
 import { migrating } from '../../src/use/migrating'
 import Clickoutside from '../../src/directives/clickoutside'
 import ElInput from '../input'
-import ElTag from '../tag'
+import { ElTag } from '../../src/components/Tag'
 import ElScrollbar from '../scrollbar'
 import ElCascaderPanel from '../cascader-panel'
 import AriaUtils from '../../src/utils/aria-utils'
