+++ packages/element3/packages/select/Select.vue
@@ -171,7 +171,7 @@ import Locale from '../../src/mixins/locale'
 import ElInput from '../input'
 import ElSelectMenu from './SelectDropdown'
 import ElOption from '../option/Option'
-import ElTag from '../tag'
+import { ElTag } from '../../src/components/Tag'
 import ElScrollbar from '../scrollbar'
 import { debounce } from 'throttle-debounce'
 import Clickoutside from '../../src/directives/clickoutside'
